fix: correct volume snapshot location config (#35)
Closes: #34 

The Velero `VolumeSnapshotLocation` does not take the same config flags
as the `BackupStorageLocation`, see

-
https://github.com/vmware-tanzu/velero-plugin-for-aws/blob/main/volumesnapshotlocation.md
-
https://github.com/vmware-tanzu/velero-plugin-for-aws/blob/main/backupstoragelocation.md

So we only need to set the `region` for the `VolumeSnapshotLocation`. 

Additionally, the PR makes the `region` config mandatory, as it's a
required config for both locations as per the links above(even for a
local deployment).

Finally, the PR introduces a check for correctly configured
`VolumeSnapshotLocation`. Now, during the backup, the test includes
`persistentvolumes` triggering Velero to validate the location.