feat: handle `upgrade_charm` event (#60)
The logic conforms to the instructions provided in the [official Velero
upgrade docs](https://velero.io/docs/v1.15/upgrade-to-1.15/)

The charm uses the newly installed binary after `juju refresh` is run to
generate and apply a list of CRDs.

After the `UpgradeCharmEvent` is handled, Juju fires the `ConfigChanged`
event, which is handled by the `reconcile` function. The triggered event
will update the images of the Deployment and NodeAgent, if present

Closes: #58